+++ docs/CHANGELOG.md
@@ -1,5 +1,10 @@
 [Home](https://tmsz-dev.github.io/Persistence/) / [Changelog](https://tmsz-dev.github.io/Persistence/CHANGELOG.html)
 
+# [2.2.0] 2021-03-25
+
+* Added [answer command](https://github.com/TMSZ-Dev/Persistence/blob/master/src/commands/challenge/answer.js)
+* Created [fs-extender.writeFileRecursive](https://github.com/TMSZ-Dev/Persistence/blob/master/src/include/fs-extender.js) to write a file and recursively make the parent directories if they don't exist
+
 # [2.1.0] 2021-03-18
 
 * Added [help command](https://github.com/TMSZ-Dev/Persistence/blob/master/src/commands/tools/help.js)